double to float

+++ Enarmet Tyveknægt/MainWindow.xaml.cs
@@ -24,7 +24,7 @@ namespace Enarmet_Tyveknægt
         public bool hhold3 = false;
 
         public int bet = 5; // Default bet mængde
-        public double balance; // Ens balance
+        public float balance; // Ens balance
         #endregion
 
         #region Object setup
